Update caches instead of invalidating them

+++ crates/sui-json-rpc/src/coin_api.rs
@@ -217,8 +217,8 @@ impl CoinReadApiServer for CoinReadApi {
             })?;
         Ok(Balance {
             coin_type: coin_type.to_string(),
-            coin_object_count: balance.num_coins,
-            total_balance: balance.balance,
+            coin_object_count: balance.num_coins as usize,
+            total_balance: balance.balance as u128,
             // note: LockedCoin is deprecated
             locked_balance: Default::default(),
         })
@@ -241,8 +241,8 @@ impl CoinReadApiServer for CoinReadApi {
             .map(|(coin_type, balance)| {
                 Balance {
                     coin_type: coin_type.to_string(),
-                    coin_object_count: balance.num_coins,
-                    total_balance: balance.balance,
+                    coin_object_count: balance.num_coins as usize,
+                    total_balance: balance.balance as u128,
                     // note: LockedCoin is deprecated
                     locked_balance: Default::default(),
                 }
